# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export type BusinessSection = typeof BusinessSection[keyof typeof BusinessSection];


export const BusinessSection = {
  catering: 'catering',
  studio: 'studio',
  beauty: 'beauty',
  supply: 'supply',
} as const;

export interface PublishedBusinessPhoto {
  section: BusinessSection;
  alt: string;
  version: number;
}

export interface AdminBusinessPhoto {
  section: BusinessSection;
  alt: string;
  smallPath: string;
  largePath: string;
  updatedAt: string;
}

export interface PublishBusinessPhotoInput {
  section: BusinessSection;
  smallPath: string;
  largePath: string;
  alt: string;
  approved: true;
}

export interface BusinessPhotoUploadRequest {
  contentType: 'image/webp';
  /**
     * @minimum 1
     * @maximum 4000000
     */
  size: number;
}

export interface BusinessPhotoUploadResponse {
  uploadUrl: string;
  objectPath: string;
}

export interface HealthStatus {
  status: string;
}