# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: photos
    description: ALCA approved business photos
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /business-photos:
    get:
      operationId: listPublishedBusinessPhotos
      tags: [photos]
      summary: Public approved photos
      responses:
        "200":
          description: Photo metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublishedBusinessPhoto"
  /business-photos/{section}/{size}:
    get:
      operationId: getBusinessPhotoImage
      tags: [photos]
      summary: Serve a published responsive image
      parameters:
        - name: section
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/BusinessSection"
        - name: size
          in: path
          required: true
          schema:
            type: string
            enum: [small, large]
      responses:
        "200":
          description: Optimized WebP image
          content:
            image/webp:
              schema:
                type: string
                format: binary
  /admin/business-photos:
    get:
      operationId: listAdminBusinessPhotos
      tags: [photos]
      summary: List published photos for the ALCA admin
      responses:
        "200":
          description: Admin photo records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminBusinessPhoto"
        "401":
          description: Sign-in required
        "403":
          description: Not authorized
    put:
      operationId: publishBusinessPhoto
      tags: [photos]
      summary: Publish an approved image after upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishBusinessPhotoInput"
      responses:
        "200":
          description: Published photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBusinessPhoto"
        "401":
          description: Sign-in required
        "403":
          description: Not authorized
  /admin/business-photos/upload-url:
    post:
      operationId: requestBusinessPhotoUpload
      tags: [photos]
      summary: Request a short-lived direct upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessPhotoUploadRequest"
      responses:
        "200":
          description: Upload URL and persistent object path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessPhotoUploadResponse"
        "401":
          description: Sign-in required
        "403":
          description: Not authorized
  /admin/business-photos/{section}:
    delete:
      operationId: unpublishBusinessPhoto
      tags: [photos]
      summary: Return a section to its labeled illustrative image
      parameters:
        - name: section
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/BusinessSection"
      responses:
        "204":
          description: Photo unpublished
        "401":
          description: Sign-in required
        "403":
          description: Not authorized
components:
  schemas:
    BusinessSection:
      type: string
      enum: [catering, studio, beauty, supply]
    PublishedBusinessPhoto:
      type: object
      required: [section, alt, version]
      properties:
        section:
          $ref: "#/components/schemas/BusinessSection"
        alt:
          type: string
        version:
          type: integer
    AdminBusinessPhoto:
      type: object
      required: [section, alt, smallPath, largePath, updatedAt]
      properties:
        section:
          $ref: "#/components/schemas/BusinessSection"
        alt:
          type: string
        smallPath:
          type: string
        largePath:
          type: string
        updatedAt:
          type: string
          format: date-time
    PublishBusinessPhotoInput:
      type: object
      required: [section, smallPath, largePath, alt, approved]
      properties:
        section:
          $ref: "#/components/schemas/BusinessSection"
        smallPath:
          type: string
        largePath:
          type: string
        alt:
          type: string
        approved:
          type: boolean
          const: true
    BusinessPhotoUploadRequest:
      type: object
      required: [contentType, size]
      properties:
        contentType:
          type: string
          const: image/webp
        size:
          type: integer
          minimum: 1
          maximum: 4000000
    BusinessPhotoUploadResponse:
      type: object
      required: [uploadUrl, objectPath]
      properties:
        uploadUrl:
          type: string
          format: uri
        objectPath:
          type: string
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status