# lib/api-zod/src/generated/types/adminBusinessPhoto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessSection } from './businessSection';

export interface AdminBusinessPhoto {
  section: BusinessSection;
  alt: string;
  smallPath: string;
  largePath: string;
  updatedAt: Date;
}