# lib/api-zod/src/generated/types/businessPhotoUploadRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessPhotoUploadRequest {
  contentType: 'image/webp';
  /**
     * @minimum 1
     * @maximum 4000000
     */
  size: number;
}