# lib/api-zod/src/generated/types/businessSection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type BusinessSection = typeof BusinessSection[keyof typeof BusinessSection];


export const BusinessSection = {
  catering: 'catering',
  studio: 'studio',
  beauty: 'beauty',
  supply: 'supply',
} as const;