# lib/api-zod/src/generated/types/publishBusinessPhotoInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessSection } from './businessSection';

export interface PublishBusinessPhotoInput {
  section: BusinessSection;
  smallPath: string;
  largePath: string;
  alt: string;
  approved: true;
}