# lib/api-zod/src/generated/types/publishedBusinessPhoto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessSection } from './businessSection';

export interface PublishedBusinessPhoto {
  section: BusinessSection;
  alt: string;
  version: number;
}